moving font function to OSS (#37619)
