+++ src/legacy/core_plugins/interpreter/public/functions/__tests__/font.js
@@ -1,16 +1,26 @@
 /*
- * Copyright Elasticsearch B.V. and/or licensed to Elasticsearch B.V. under one
- * or more contributor license agreements. Licensed under the Elastic License;
- * you may not use this file except in compliance with the Elastic License.
+ * Licensed to Elasticsearch B.V. under one or more contributor
+ * license agreements. See the NOTICE file distributed with
+ * this work for additional information regarding copyright
+ * ownership. Elasticsearch B.V. licenses this file to you under
+ * the Apache License, Version 2.0 (the "License"); you may
+ * not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *    http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing,
+ * software distributed under the License is distributed on an
+ * "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
+ * KIND, either express or implied.  See the License for the
+ * specific language governing permissions and limitations
+ * under the License.
  */
 
 import expect from '@kbn/expect';
-import { openSans } from '../../../../common/lib/fonts';
+import { openSans } from '../../../common/lib/fonts';
 import { font } from '../font';
-import { functionWrapper } from '../../../../__tests__/helpers/function_wrapper';
-import { getFunctionErrors } from '../../../strings';
-
-const errors = getFunctionErrors().font;
+import { functionWrapper } from '../../../test_helpers';
 
 describe('font', () => {
   const fn = functionWrapper(font);
@@ -50,7 +60,7 @@ describe('font', () => {
 
       it('defaults to 1', () => {
         const result = fn(null);
-        expect(result.spec).to.have.property('lineHeight', 1);
+        expect(result.spec).to.have.property('lineHeight', '1');
         expect(result.css).to.contain('line-height:1');
       });
     });
@@ -100,16 +110,14 @@ describe('font', () => {
         expect(result.css).to.contain('font-weight:400');
       });
 
-      it("defaults to 'normal'", () => {
+      it('defaults to \'normal\'', () => {
         const result = fn(null);
         expect(result.spec).to.have.property('fontWeight', 'normal');
         expect(result.css).to.contain('font-weight:normal');
       });
 
       it('throws when provided an invalid weight', () => {
-        expect(() => fn(null, { weight: 'foo' })).to.throwException(
-          new RegExp(errors.invalidFontWeight('foo').message)
-        );
+        expect(() => fn(null, { weight: 'foo' })).to.throwException();
       });
     });
 
@@ -177,7 +185,7 @@ describe('font', () => {
       it('throws when provided an invalid alignment', () => {
         expect(fn)
           .withArgs(null, { align: 'foo' })
-          .to.throwException(errors.invalidTextAlignment('foo').message);
+          .to.throwException();
       });
     });
   });
