+++ x-pack/legacy/plugins/canvas/canvas_plugin_src/strings/functions/function_errors.ts
@@ -10,7 +10,6 @@ import { errors as compare } from './compare';
 import { errors as containerStyle } from './containerStyle';
 import { errors as csv } from './csv';
 import { errors as date } from './date';
-import { errors as font } from './font';
 import { errors as getCell } from './getCell';
 import { errors as image } from './image';
 import { errors as math } from './math';
@@ -28,7 +27,6 @@ export const getFunctionErrors = () => ({
   containerStyle,
   csv,
   date,
-  font,
   getCell,
   image,
   math,
